work on reparam

+++ src/constants.hpp
@@ -44,10 +44,10 @@ static const name& REX_ACCOUNT  = EOSIO;
 
 // constants in %
 
-const uint8_t CR = 150; /// minimal collateral ratio
-const uint8_t LF = 10;  /// liquidation fee
-const uint8_t RF = 1;   /// redemption fee
-const uint8_t SP = 20;  /// part of taxes for scruge
-const uint8_t SR = 80;   /// savings ratio
-const uint8_t IR = 20;   /// insurance ratio
+const uint64_t CR = 150; /// minimal collateral ratio
+const uint64_t LF = 10;  /// liquidation fee
+const uint64_t RF = 1;   /// redemption fee
+const uint64_t SP = 20;  /// part of taxes for scruge
+const uint64_t SR = 80;   /// savings ratio
+const uint64_t IR = 20;   /// insurance ratio
 const double AR = 0.05;   /// annual interest rate
